// frontend/src/types/newUser.ts
import NewCourse from "./newCourse";

// called NewUser because we used to have a type called User in a1, and this replaces it
// not to be confused with new user that just signed up

interface NewUser {
	id?: number; // optional because it is autogenerated by backend
	firstName: string;
	lastName: string;
	email: string;
	role: "Lecturer" | "Candidate" | "Admin";
	password: string;
	course?: NewCourse;
	createdAt?: Date;
	updatedAt?: Date;
}

export default NewUser;